requirements.txt
asgiref             3.8.1
black               24.4.2
click               8.1.7
colorama            0.4.6
Django              4.2
djangorestframework 3.15.2
mypy-extensions     1.0.0
packaging           24.1
pathspec            0.12.1
pillow              10.4.0
pip                 24.1.1
platformdirs        4.2.2
psycopg2-binary     2.9.9
python-dotenv       1.0.1
setuptools          70.0.0
sqlparse            0.5.0
tzdata              2024.1
wheel               0.43.0
django-celery-beat            2.6.0


Django~=5.0.4
python-dotenv~=1.0.1
djangorestframework~=3.15.2
django-filter~=24.2
djangorestframework-simplejwt 5.3.1
celery~=5.4.0
drf_yasg
stripe~=10.3.0
config~=0.5.1
pytz~=2024.1